# openapi.yaml
openapi: 3.1.0
info:
  title: Dejima API
  version: "0.1"
  x-api-version: 1
  description: >
    HTTP/WebSocket control surface for the Dejima daemon (`dejimad`). This is the
    same API the CLI/TUI use and the contract third-party apps build on.

    Status: Dejima is in alpha (0.x). The `v1/` path prefix is intended to be
    stable, but fields may change until `v1.0.0`; an integer `api_version`
    (advertised on `/v1/overview`) tracks the client/daemon contract separately
    from the semver tag. This spec is hand-maintained from the route table in
    `internal/api/server.go` and the request/response shapes in
    `internal/api/types.go` — those are the source of truth and the spec may
    drift slightly between releases.

    WebSocket endpoints (PTY session streams) are listed for discoverability but
    cannot be fully modeled in OpenAPI; see the per-path notes and the
    `SessionEnvelope` schema for the wire protocol.
servers:
  - url: http://{host}
    description: Daemon over the tailnet (DEJIMA_HOST, e.g. 100.84.12.7:7273)
    variables:
      host:
        default: 127.0.0.1:7273
security:
  - bearerToken: []
  - {}   # operator listeners (unix socket / tailnet) carry no per-request token
...
paths:
  /v1/islands:
    get:
      tags: [islands]
      summary: List all islands
      operationId: listIslands
      responses:
        "200":
          description: All islands, most-recently-used first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/IslandInfo" }
    post:
      tags: [islands]
      summary: Create (provision) an island
      operationId: createIsland
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateIslandRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateIslandResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: An island with this name already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /v1/activity:
    get:
      tags: [activity]
      summary: Team activity feed — curated "who launched what, which agent did what"
      description: >
        A read-only, human-rendered timeline projected over the audit ledger:
        operator/account actions (from `api.request` records, when `--audit` is
        on) plus always-on brokered agent↔host activity (port/capability/MCP),
        enriched with each island's owner. Newest-first. A scoped token only sees
        activity for islands in its scope. `audit_enabled` flags whether the
        operational log is on (when off, only brokered records appear).
      operationId: activity
      parameters:
        - name: actor
          in: query
          schema: { type: string }
          description: Only items by this actor (token label / operator / island:<name>)
        - name: island
          in: query
          schema: { type: string }
        - name: owner
          in: query
          schema: { type: string }
          description: Only items for islands with this owner label
        - name: kind
          in: query
          schema: { type: string, enum: [lifecycle, broker, system] }
        - name: decision
          in: query
          schema: { type: string, enum: [allowed, denied] }
        - name: since
          in: query
          schema: { type: string, format: date-time }
          description: Only items at/after this RFC3339 time
        - name: until
          in: query
          schema: { type: string, format: date-time }
          description: Only items at/before this RFC3339 time
        - name: limit
          in: query
          schema: { type: integer }
          description: Cap the number of items returned (after filtering)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActivityResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /v1/audit:
    get:
      tags: [daemon]
      summary: The audit ledger (brokered ops + api.request + lifecycle) with filtering
      description: >
        Returns the hash-chained audit ledger plus a whole-chain verification
        verdict. Filters compose; `limit` tails the filtered result. With
        `format=jsonl` or `format=csv` the body is a plain-text export of the
        filtered entries instead of the JSON envelope.
      operationId: audit
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
          description: Tail the last N entries after filtering (the full chain is still verified)
        - name: since
          in: query
          schema: { type: string, format: date-time }
          description: Only entries at/after this RFC3339 time
        - name: until
          in: query
          schema: { type: string, format: date-time }
          description: Only entries at/before this RFC3339 time
        - name: island
          in: query
          schema: { type: string }
          description: Only entries for this island
        - name: type
          in: query
          schema: { type: string }
          description: "Only entries whose type has this prefix (e.g. port., api.request)"
        - name: actor
          in: query
          schema: { type: string }
          description: Only entries by this actor (identity subject)
        - name: decision
          in: query
          schema: { type: string, enum: [allowed, denied] }
        - name: format
          in: query
          schema: { type: string, enum: [json, jsonl, csv], default: json }
          description: "json (default) returns AuditResponse; jsonl/csv stream a text export"
      responses:
        "200":
          description: "AuditResponse (format=json), or a text/plain export (jsonl/csv)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditResponse" }
            text/plain:
              schema: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /v1/internal/agent-event:
    post:
      tags: [daemon]
      summary: "Internal — agent telemetry sink (notify-hook shims). Not for app use."
      description: >
        Posted by the in-island notify hooks. A token-authenticated caller may
        only emit for its own island (the body's island is overridden by the
        token's island). Listed for completeness; third-party apps consume events
        via webhooks, not by posting here.
      operationId: agentEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AgentEventRequest" }
      responses:
        "202": { description: Accepted }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: Per-island bearer token (DEJIMA_TOKEN) for the autonomy path.
...
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
    CreateIslandRequest:
      type: object
      required: [repo]
      properties:
        name: { type: string, description: optional; derived from repo if empty }
        repo: { type: string, description: git URL or local path }
        agent: { type: string, default: claude-code }
        image: { type: string, default: "dejima/island:latest" }
        cmd: { type: string, description: "entrypoint for agent=headless (required then; ignored otherwise)" }
        resources: { $ref: "#/components/schemas/Resources" }
        agents:
          type: array
          description: "Seed a multi-agent island; element 0 is the primary."
          items: { $ref: "#/components/schemas/AgentSpecRequest" }
        role: { type: string, description: "\"\" (work island) or \"home\" (assistant brain; must be headless)" }
        github_identity: { type: string, description: which daemon GitHub identity to clone/push as }
        seed_path: { type: string, description: host path bind-mounted read-only as the clone source (local daemon only) }
        owner: { type: string }
        tags: { type: object, additionalProperties: { type: string } }
    CreateIslandResponse:
      allOf:
        - $ref: "#/components/schemas/IslandInfo"
        - type: object
          properties:
            token:
              type: string
              description: only present on a token-authenticated (Home-Island spawn) create
...
    AgentSpecRequest:
      type: object
      properties:
        type: { type: string, description: "agent handler — claude-code, codex, shell, openclaw, headless, …" }
        label: { type: string }
        cmd: { type: string, description: required for a generic headless type }
        provider: { type: string, description: names a daemon provider credential }
        model: { type: string, description: "\"provider/model\" string" }
...
    Resources:
      type: object
      properties:
        memory: { type: string }
        cpus: { type: string }
        disk: { type: string }
        oom_priority: { type: [integer, "null"] }
...
    IslandInfo:
      type: object
      properties:
        name: { type: string }
        title: { type: string }
        repo: { type: string }
        agent: { type: string }
        image: { type: string }
        cmd: { type: string }
        role: { type: string }
        owner: { type: string }
        tags: { type: object, additionalProperties: { type: string } }
        state: { type: string, description: desired state from config }
        container: { type: string, description: observed runtime status }
        created_at: { type: string, format: date-time }
        last_used_at: { type: string, format: date-time }
        attached:
          type: array
          items: { $ref: "#/components/schemas/PresenceEntry" }
        stats: { $ref: "#/components/schemas/IslandStats" }
        agent_state: { $ref: "#/components/schemas/AgentStateInfo" }
        git: { $ref: "#/components/schemas/GitInfo" }
        health: { $ref: "#/components/schemas/IslandHealth" }
        disk: { $ref: "#/components/schemas/IslandDisk" }
        resources: { $ref: "#/components/schemas/Resources" }
        agents:
          type: array
          items: { $ref: "#/components/schemas/AgentInfo" }
    AgentInfo:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        label: { type: string }
        tmux: { type: string }
        branch: { type: string }
        worktree: { type: string }
        attachable: { type: boolean }
        created_at: { type: string, format: date-time }
        state: { type: string, description: "session liveness: running | stopped | exited (detail only)" }
        agent_state: { $ref: "#/components/schemas/AgentStateInfo" }
        attached:
          type: array
          items: { $ref: "#/components/schemas/PresenceEntry" }
        restarts: { type: integer }
        error: { type: string, description: last orchestration failure }
        error_at: { type: string, format: date-time }
        provider: { type: string }
        model: { type: string }
        provider_key_set: { type: boolean }
        auth_state: { type: string, description: "\"missing-provider-auth\" or \"\" (ready)" }
    PresenceEntry:
      type: object
      properties:
        label: { type: string }
        joined_at: { type: string, format: date-time }
    AgentStateInfo:
      type: object
      properties:
        latest: { type: string, description: "e.g. waiting-for-input, task-complete, error" }
        updated_at: { type: string, format: date-time }
    IslandStats:
      type: object
      properties:
        memory_usage_bytes: { type: integer }
        memory_limit_bytes: { type: integer }
        cpu_percent: { type: number }
    GitInfo:
      type: object
      properties:
        branch: { type: string }
        clean: { type: boolean }
        ahead: { type: integer }
        behind: { type: integer }
        dirty_files: { type: integer }
    IslandHealth:
      type: object
      properties:
        oom_killed: { type: boolean }
        restart_count: { type: integer }
        exit_code: { type: integer }
    IslandDisk:
      type: object
      properties:
        workspace_bytes: { type: integer }
        home_bytes: { type: integer }
        total_bytes: { type: integer }
...
    AgentEventRequest:
      type: object
      required: [type]
      properties:
        island: { type: string }
        agent: { type: string }
        type: { type: string }
        payload: { type: object, additionalProperties: true }
...
    LedgerEntry:
      type: object
      properties:
        seq: { type: integer }
        time: { type: string, format: date-time }
        type: { type: string, description: "port.grant | port.revoke | trade.read | trade.write | capability.* | mcp.* | api.request | <lifecycle event>" }
        island: { type: string }
        agent: { type: string }
        scope: { type: string }
        path: { type: string, description: "host/scope path (trades) or request path (api.request)" }
        mode: { type: string }
        bytes: { type: integer }
        sha256: { type: string }
        decision: { type: string, description: allowed | denied }
        detail: { type: string }
        method: { type: string, description: "HTTP method (api.request)" }
        status: { type: integer, description: "HTTP status code (api.request)" }
        actor: { type: string, description: "who made the request (identity; filled by the auth layer)" }
        role: { type: string, description: "the actor's role, when known" }
        prev: { type: string, description: chain value of the previous entry }
        chain: { type: string, description: chain value of this entry }
    AuditResponse:
      type: object
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/LedgerEntry" }
        total: { type: integer, description: entries in the whole ledger (before filtering) }
        returned: { type: integer, description: entries returned after filter + limit }
        verified: { type: boolean, description: whole-chain hash verification result }
        error: { type: string, description: chain-verification failure detail }
    ActivityItem:
      type: object
      properties:
        seq: { type: integer }
        time: { type: string, format: date-time }
        actor: { type: string, description: "who acted: token label / operator / island:<name>" }
        role: { type: string }
        island: { type: string }
        owner: { type: string, description: that island's owner label (enriched) }
        kind: { type: string, enum: [lifecycle, broker, system] }
        summary: { type: string, description: human-readable one-liner }
        decision: { type: string, description: allowed | denied }
    ActivityResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ActivityItem" }
        returned: { type: integer }
        audit_enabled: { type: boolean, description: false ⇒ feed carries only always-on brokered records }
...